对应后端将学工号改为Long进行了修改
如题

+++ java-fx/src/main/java/com/teach/javafx/models/DO/Student.java
@@ -27,7 +27,7 @@ import java.io.Serializable;
 public class Student implements Serializable {
     @Id
 //    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer studentId;
+    private Long studentId;
 
     @OneToOne
     @JoinColumn(name = "person_id")
@@ -38,7 +38,7 @@ public class Student implements Serializable {
     private String className;
 
     public Student(StudentInfo studentInfo){
-        studentId= Integer.valueOf(studentInfo.getNumber());
+        studentId= studentInfo.getNumber();
         major=studentInfo.getMajor();
         className=studentInfo.getClassName();
     }
